+++ src/services/weatherService.js
@@ -25,9 +25,9 @@ export async function getMeteo(latitude, longitude) {
     // On questionne l'API pour qu'elle fournisse les info météo de la ville ayant les coordonnées renseignées
     const response = await fetch(
         `${METEO_URL}?latitude=${latitude}&longitude=${longitude}`
-        + `&current=temperature_2m,apparent_temperature,windspeed_10m,winddirection_10m,relativehumidity_2m,weathercode`
-        + `&daily=temperature_2m_max,temperature_2m_min,weathercode`
-        + `&hourly=weathercode,temperature_2m`
+        + `&current=temperature_2m,apparent_temperature,windspeed_10m,winddirection_10m,relativehumidity_2m,weathercode,precipitation_probability`
+        + `&daily=temperature_2m_max,temperature_2m_min,weathercode,precipitation_probability_mean`
+        + `&hourly=weathercode,temperature_2m,precipitation_probability`
         + `&timezone=auto`
     )
     const data = await response.json()